requirements.txt
celery==4.3.0
Django==2.2.6
djangorestframework==3.10.3
djoser==2.0.3
redis==3.3.8
requests==2.22.0